move network methods into a separate class
